Switch from system prompt to instructions
Instructions for the current agent are used even when an explicit
`message_history` is provided to `Agent.run` [1]. This is relevant when
using functionality such as `Agent.to_ag_ui()`, where the framework
maintains a message history that is passed in to the agent and does not
include the agent's system prompt. The Pydantic AI docs suggest using
`instructions` over `system_prompt` unless there is a specific reason to
use the latter.

[1]: https://ai.pydantic.dev/agents/#instructions